Kwarg cleanup
Continued implementing better kwarg saving, loading and printing. Fixed stuff with units, so that they are handled completely internally within the Simulation object, such that they are in real units when printed and saved but non-dimensionalized within the Simulation.

Also added default_kwargs.json file, for easy loading and editing.